Home Screen updates
-Added radio icon to replace navigation button on home screen
-Cleaned up App files by removing excess code
-
